# Less tagging in asmcomp-optimized switches