--- a/asmcomp/cmm_helpers.mli
+++ b/asmcomp/cmm_helpers.mli
@@ -538,10 +538,12 @@ val bigstring_set :
 (** [transl_isout h arg dbg] *)
 val transl_isout : expression -> expression -> Debuginfo.t -> expression
 
+type switch_arg = Tagged of expression | Untagged of expression
 (** [make_switch arg cases actions dbg] : Generate a Cswitch construct,
-    or optimize as a static table lookup when possible. *)
+    or optimize as a static table lookup when possible.
+*)
 val make_switch :
-  expression -> int array -> (expression * Debuginfo.t) array -> Debuginfo.t ->
+  switch_arg -> int array -> (expression * Debuginfo.t) array -> Debuginfo.t ->
   expression
 
 (** [transl_int_switch loc arg low high cases default] *)