@gasche 2024-10-22
The testsuite-stability problem is on my mind, and I was vaguely thinking of providing more options to control it better. (I thought of writing a post-processing step to remove the noisy parts away, of improving the implenetation of `-dno-unique-ids`, and/or of adding an option to show the cmm only for a specific option. I wasn't motivated enough to do any of it yet.) I agree to follow your suggestion of removing the test after approval, although I would note that in practice these tests are useful to me over time, and can protect from code-quality regressions. (I have committed (in both senses of the term) tests that store -dlambda output, and when they evolve because code-generation changes I usually find the results relevant and interesting.) This is a problem that will show up with other PRs, for example the exact same how-to-test-the-cmm problem occurs with atomic record fields.

> You're removing the possibility of making affine functions out of switches with non-tagged constants.

No, I think that this possibility was not present in trunk already. If you look at `extract_uconstant`, it returns None for constants that have least bit 0, so neither tablification nor affinification can apply.

> It's annoying that we have to wait until Cmm to detect affine switches on OCaml integers.

We could decide to move the optimization to the Lambda level, and/or we could package it as a functor like the Switch module to use it in asmcomp/ and lambda/ without too much code duplication, or maybe have it in both closure and flambda if that makes sense. (But I won't claim that thinking about Switch in both contexts is my favorite part of the code either.) I would propose to consider this outside the scope of this PR.