@gasche 2024-10-22
One thing that I forgot to mention: these optimizations (tables and affine transformations) are carefully tested by testsuite/tests/basic/switch_opts.ml, which exercises both code paths (on tagged and untagged integers). I know because intermediate versions of the PR had correctness bugs and they were reliably caught by this testsuite. I believe that that test is a good argument for correctness of the new version; the test that I added myself is there to also check that it actually improves the generated code (slightly).