@gasche 2024-10-24
Thanks! I removed (with regret) the -dcmm tests, as we discussed.